pyproject.toml
[project]
name = "dynamic-network-slicing-live-events"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "databricks-connect>=15.4.5,<=15.4.5",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "mlflow>=2.10.0",
    "databricks-sdk>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
